Fix: Return role name in map correctly to fix welcome embed list
Javascript automatically returns whatever is after the => if the
function is within a single line. With the brackets there, it'll
treat whatever is within the brackets as a function instead, and
expect it to contain a "return"; as a result, each loop of the current
map returns nothing and we have whitespace joined.