Fix: Check existing roles correctly so we don't re-add existing roles
The guild.roles.fetch promise resolves with a Collection class of their
Role object. Although they support has instead of hasOwnProperty, they
use the ID for keys instead of the role names that we check here.

Instead, we'll use their .find helper to make sure we can't find
a role with the same name before trying to create it.