Relocate registerGuildCommands to guildCreate
Now that we know how to add a bot to a discord server with the desired
scope/permissions, we no longer need to explicitly ask an admin to
add the server to the bot after it first joins.

This means the guildCreate handler can now immediately try to register
our commands instead of waiting for the button press caught by
interactionCreate.