Update README to reflect new steps
The README now suggests the additional necessary scope for the
authorization URL, and no longer recommends the OAuth2 configuration
for setting up the bot for testing.
The README now also separates the Juno testnet token steps from
the info actually necessary for getting started. The steps now
also reflect the disappearance of the #verify-here room.