Relocate registerGuildCommands to guildCreate (#43)
* Relocate registerGuildCommands to guildCreate

Now that we know how to add a bot to a discord server with the desired
scope/permissions, we no longer need to explicitly ask an admin to
add the server to the bot after it first joins.

This means the guildCreate handler can now immediately try to register
our commands instead of waiting for the button press caught by
interactionCreate.

* Update README to reflect new steps

The README now suggests the additional necessary scope for the
authorization URL, and no longer recommends the OAuth2 configuration
for setting up the bot for testing.
The README now also separates the Juno testnet token steps from
the info actually necessary for getting started. The steps now
also reflect the disappearance of the #verify-here room.