Overhaul Wizard system with Command Chain
This refactoring converts the Wizard Chain functionality into
an option for every command configured for the bot. Using a pattern
similar to express and koa middleware, each new command step can
now expect the same new params as commands: req, res, ctx and next.

The last next function effectively allows any command to kick off
a chain by calling next with a function that resolves what command
should be next if an interaction satisfactorily matches the conditions
for the next command to be kicked off.