Fix: Update handleCW20Entry for new testnet URL
The URL for daodao's testnet is now testnet.daodao.zone, which
breaks our handleCW20Entry and checkForDAODAODAO logic, which was
expecting the daodao URL to always look a certain way.

Both of these cases now use regexes instead. This also updates
the functions in token to throw their errors again, and rely on the
step calling them to call res.error more correctly in the catch
function.