✨ Add /token-rule remove command chain functionality
Admin users can now call /starry token-rule remove to see a list
of token rule roles that they can ask to remove via the bot.
When a selection is made, the removeVerify step replicates the
farewell confirmation check with 2 buttons that take us to
removeConfirmation and removeRejection respectively.

This also adds:
- New roleGet functionality in db.js to only get the data for
  as single role in a single guild
- Small fix to make the done message reply to the last interaction
  in the chain, which prevents the last buttons from thinking
  they never finished being interacted with.