Reregister commands for guilds on bot restart as needed
The bot currently only registers guild commands on guildCreate,
which is the event used when a bot is first added to a guild. This
inadvertently meant that users of starrybot could not get any of
the recent slash command updates unless they kicked and re-added
the bot.

Now, when the bot has successfully started, we'll try and see if
we can helpfully re-register guild commands for any of the guilds
registered with our client. If there are no differences detected
between their registered commands and our latest starry commands,
nothing will happen.