Specify interaction targets for correct reply target
Rather than trying to guesstimate which discord message wants the
reply sent by a wizard step, we'll have the event handlers pass
the desired reply target in when they continue the wizard, since
they know what type of action they're handling.

This also fixes a bug caused by our reliance on the existence of the
"interaction" vs. "reaction" props to determine which to use for
a reply; any button or user input message after an emoji reaction
message would respond to the incorrect message for the duration
of the wizard.