Bugfix for immortal wizards
tl;dr Javascript changes what "this" refers to when functions are
passed along in certain ways - we used the wrong way for trying
to timeoue + end our wizards.

Our this.end function in the wizard class currently requires "this"
to be set to the Wizard instance so that it can correctly reference
the parent wizard for deleting itself from the global map - however,
the way it was being passed in caused "this" to always be undefined,
and wizards were living on forever.

This caused a rare bug that appeared only if a user timed out on the
last step of a wizard that was awaiting user input (every other step
would subsequently "correct" the issue by updating the map with the
next "next" function).