Only continue wizard on correct prompt type
As part of calling `next`, the wizard now tracks what type
of prompt the next step is waiting for. Each discord event
handler now informs the wizard on continue which prompt
type it believes it has (reaction, button, input), and now
the wizard will only continue if the prompt type matches
the one we think we're waiting for.

This means that typing a message when the wizard is expecting
a button or reaction won't crash it, and reacting when it's
expecting a button or input won't crash it, etc. etc.