Upgrade discordjs from v13 to v14
Discord no longer wants bots to read message contents for security
reasons, which conflicts with the "installation wizard-like" UX
pattern we had previously landed on for the /starry token-rule
commands. Their recommendation is to replace that with their Modal
and Select dropdown components instead, which come with v14.

This commit upgrades the node version to the minimum required for
v14 and bumps up the discordjs package itself.