Minor adjustments for various command types
Modal commands must always be first, which means they cannot be
deferred. Given that, we update our sequence of operations to
now be:

1. Immediately reply an error if necessary
2. Create a modal if necessary
3. Defer the reply if possible (i.e. deferReply exists), or
4. Finally create a normal reply

This change means that there's now a case where deferReply may
exist on the interaction but was not actually used because the
command prompt type was a modal instead. In order to avoid a
breaking editReply call, we now check if the interaction was
deferred instead.

This commit additionally adds a new feature for buttons, to
allow buttons to go to either different steps or the same steps
(with unique customIds/values configured). This will be used
in subsequent commits.