Deprecate message content use, reduce permissions
With all uses of the 'input' command prompt type refactored in
previous commits, we can finally remove that feature entirely! This
means:

- The 'input' type is no longer valid in the config
- The 'messageCreate' event is no longer interesting to us
- The bot no longer needs any Privileged Gateway Intents configured
  (in production or in development)

In addition, we now have a clear understanding of exactly which
permissions are necessary for the bot to function. The README has
been updated with our new recommended URL, so discord guilds no longer
need to install it with the intimidating admin scope, and the website
with the installation button will be updated shortly.