Call deferReply earlier, add helper for responses
We were running into an "Unknown Interaction" error when a getConfig
function took more than 3 seconds to finish. We originally intended
for the deferReply feature to address this, but we weren't actually
deferring until after getConfig had finished.

In this commit, all steps will now defer the reply before getConfig
is called unless otherwise configured (i.e. deferReply: false) on the
command. The only steps that don't want the reply deferred today are
the ones using the modal type, as the modal must be created before
anything else happens.

There is now also a "respond" helper function for all command prompt
types to use, so we don't have to keep checking for the deferred
status + try/catch correctly each time.