Support legacy + updated DAO tokens rules
In the latest version of Daodao, the get_config endpoint no
longer works for fetching the governance token and staking contract.
However, Daos that not yet updated to the latest version are also
still functional.

This commit tries to fetch the governance token and staking contract
from the DAO address via the updated queries first. If that fails,
we'll still try using the legacy endpoint before finally giving up.