Refactoring commands to avoid user inputs (#115)
* Support select and modal command types

Discord.js doesn't like letting bots having access to message
contents unless absolutely necessary. Our installation wizard-like
UX is not considered "necessary" by their standards, so we'll
replace them with their recommendation interaction patterns
instead: select dropdowns and modal forms.

This commit adds support for these two types without yet removing
the emoji or input types so that this branch can support an
incremental roll-out.

* Minor adjustments for various command types

Modal commands must always be first, which means they cannot be
deferred. Given that, we update our sequence of operations to
now be:

1. Immediately reply an error if necessary
2. Create a modal if necessary
3. Defer the reply if possible (i.e. deferReply exists), or
4. Finally create a normal reply

This change means that there's now a case where deferReply may
exist on the interaction but was not actually used because the
command prompt type was a modal instead. In order to avoid a
breaking editReply call, we now check if the interaction was
deferred instead.

This commit additionally adds a new feature for buttons, to
allow buttons to go to either different steps or the same steps
(with unique customIds/values configured). This will be used
in subsequent commits.

* Refactor token-rule remove to not use user input

- The first step of /starry token-rule remove now provides a select
  dropdown instead of requiring the user to type in the exact name
  of the role they'd like to remove.
- The removeConfirmation and removeRejection steps have been
  consolidated into a single removeRole step. The removeVerify step
  will always navigate to removeRole now, as removeRole will check
  which button was actually pressed by the user.

* Refactor token-rule edit to not use user input

This commit drastically simplifies the flow for this command
from what it was before:

   index -> editCheck -> editRoleName -> handleRoleNameEdit
                         editRoleAmount -> handleRoleAmountEdit
                         editRoleStakedOnly -> handleRoleStakedOnly
into the following:

   index -> editRole -> editRoleStakedOnly -> handleRoleEdit

- index now displays a select dropdown instead of requiring the
  user to type in the exact name of the role to edit
- editRole now opens a modal instead of requiring the user to edit
  the name or amount only, in separate input flows
- editRoleStakedOnly now validates the data from the modal before
  providing the buttons for staked vs not staked
- handleRoleEdit now handles updating the token rule with all of
  the requested changes (name, amount, staked or not)

* Refactor token-rule add to not use user input

This commit drastically simplies the flow for this command
from what it was before:

  index -> addNativeToken -> nativeTokenJUNO -> promptTokenName -> handleStakedOnlyYes
                                                                   handleStakedOnlyNo
                             nativeTokenSTARS -> promptTokenName -> handleStakedOnlyYes
                                                                    handleStakedOnlyNo
                             nativeTokenSuggestion
        -> addCW20 -> needsCW20 -> promptTokenName -> handleStakedOnlyYes
                                                   -> handleStakedOnlyNo
                      hasCW20 -> promptTokenName -> handleStakedOnlyYes
                                                    handleStakedOnlyNo
                      daoDao -> promptTokenName -> handleStakedOnlyYes
                                                   handleStakedOnlyNo
        -> addCW721 -> hasCW721 -> promptTokenName -> handleStakedOnlyYes
                                                   -> handleStakedOnlyNo
                       stargaze -> promptTokenName -> handleStakedOnlyYes
                                                   -> handleStakedOnlyNo
        -> explainTokenTypes -> //recursive back to index

into the following:

  index -> addNativeToken -> promptNativeToken -> createTokenRule
        -> addCW20 -> promptCW20 -> createTokenRule
        -> addCW721 -> promptCW721 -> createTokenRule

because dear god

- index now displays a select dropdown of our options instead of
  prompting for emoji reactions as it did before. This also means that
  the information previously provided by explainTokenTypes has been
  shifted into the description.
- The various add* steps now all provide select dropdowns instead of
  emoji reactions, and thehe respective prompt* steps now all open a
  form modal with helper text personalized to their previous selection.
- The final createTokenRule validates all of the previously requested
  configurations before creating the token role. This means that this
  final step also includes all the selection-specific handling from
  before, i.e. daoDao, stargaze, etc.

* Deprecate message content use, reduce permissions

With all uses of the 'input' command prompt type refactored in
previous commits, we can finally remove that feature entirely! This
means:

- The 'input' type is no longer valid in the config
- The 'messageCreate' event is no longer interesting to us
- The bot no longer needs any Privileged Gateway Intents configured
  (in production or in development)

In addition, we now have a clear understanding of exactly which
permissions are necessary for the bot to function. The README has
been updated with our new recommended URL, so discord guilds no longer
need to install it with the intimidating admin scope, and the website
with the installation button will be updated shortly.

* Update README to not have outdated node version

* Remove deferred message handling temporarily

Something about the deferReply handling is causing us to be unable
to reply to the ModalSubmitInteraction consistently. We're removing
it to debug before putting it back in.

* Call deferReply earlier, add helper for responses

We were running into an "Unknown Interaction" error when a getConfig
function took more than 3 seconds to finish. We originally intended
for the deferReply feature to address this, but we weren't actually
deferring until after getConfig had finished.

In this commit, all steps will now defer the reply before getConfig
is called unless otherwise configured (i.e. deferReply: false) on the
command. The only steps that don't want the reply deferred today are
the ones using the modal type, as the modal must be created before
anything else happens.

There is now also a "respond" helper function for all command prompt
types to use, so we don't have to keep checking for the deferred
status + try/catch correctly each time.

* Fix: Respond with error messages correctly

* Fix STARS network value so the role works again

* Fix for token amount-only edits

* Restore embeds configured on button prompt types

* Support legacy + updated DAO tokens rules

In the latest version of Daodao, the get_config endpoint no
longer works for fetching the governance token and staking contract.
However, Daos that not yet updated to the latest version are also
still functional.

This commit tries to fetch the governance token and staking contract
from the DAO address via the updated queries first. If that fails,
we'll still try using the legacy endpoint before finally giving up.

* Prevent bot from crashing from getConfig bugs