Support NFT role amounts, fix empty role bug (#117)
Previously we did not officially support CW721 token roles that
checked for multiple NFTs. This has turned out to be a surprisingly
popular usecase so it's time to make it official :D
This means:
- On token-rule add, the amount will always be visible now
- On token-rule edit, we always allow amount to be editable, and
  properly validate the amount as the steps progress.
- The calculation for multiple NFTs should work now that we've
  also manually updated the production NFT token-rules to all
  have the decimals set to 1 instead of null.

Also make sure no one can create a role name that's just
whitespace because that causes problems.